refactor(cli): share one age formatter across the CLIs
The relative-age formatter existed as five byte-identical copies, each with its
own copy-pasted tests, and a sixth better implementation had just landed in the
readiness CLI.

Keep one implementation in the shared CLI library and let every consumer call
it. The rendered age loses its trailing "ago", which the column headers already
convey, and gains the shorter, capped form of the surviving implementation.

Drop the dependency the migrated crates no longer need.
