+++ cli/modules/gateway/src/main.rs
@@ -1,5 +1,7 @@
 //! CLI for the YANET gateway service registry.
 
+use std::time::SystemTime;
+
 use clap::{ArgAction, CommandFactory, Parser};
 use clap_complete::CompleteEnv;
 use tabled::{
@@ -14,6 +16,7 @@ use tonic::codec::CompressionEncoding;
 use ync::{
     client::{ConnectionArgs, LayeredChannel, Service},
     errors::Error,
+    humanfmt,
     output::{self, CommonFormat},
 };
 use ynpb::pb::{BackendKind, ListServicesRequest, RegisteredBackend, gateway_client::GatewayClient};
@@ -156,7 +159,9 @@ fn kind_display(kind: BackendKind) -> String {
 fn last_seen_cell(kind: BackendKind, ts: Option<&prost_types::Timestamp>) -> String {
     match kind {
         BackendKind::Builtin | BackendKind::InProcess => "\u{2014}".to_string(),
-        BackendKind::External | BackendKind::Unspecified => format_age(ts),
+        BackendKind::External | BackendKind::Unspecified => {
+            humanfmt::format_age(ts, SystemTime::now()).unwrap_or_else(|| "-".to_string())
+        }
     }
 }
 
@@ -177,89 +182,9 @@ fn render_table(rows: &[ServiceRow]) {
     println!("{table}");
 }
 
-/// Formats a `prost_types::Timestamp` as a human-readable relative age.
-///
-/// Returns `"-"` when `ts` is `None` or the zero sentinel
-/// (`seconds == 0 && nanos == 0`). Otherwise formats as `Xs ago`,
-/// `XmYs ago`, or `XhYm ago` depending on magnitude.
-pub fn format_age(ts: Option<&prost_types::Timestamp>) -> String {
-    let ts = match ts {
-        Some(ts) if ts.seconds != 0 || ts.nanos != 0 => ts,
-        _ => return "-".to_string(),
-    };
-
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default();
-
-    let ts_secs = ts.seconds.max(0) as u64;
-    let now_secs = now.as_secs();
-
-    let secs = now_secs.saturating_sub(ts_secs);
-
-    if secs < 60 {
-        format!("{secs}s ago")
-    } else if secs < 3600 {
-        let minutes = secs / 60;
-        let remainder = secs % 60;
-        format!("{minutes}m{remainder}s ago")
-    } else {
-        let hours = secs / 3600;
-        let minutes = (secs % 3600) / 60;
-        format!("{hours}h{minutes}m ago")
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
-
-    #[test]
-    fn format_age_none_returns_dash() {
-        assert_eq!("-", format_age(None));
-    }
-
-    #[test]
-    fn format_age_zero_sentinel_returns_dash() {
-        let ts = prost_types::Timestamp { seconds: 0, nanos: 0 };
-        assert_eq!("-", format_age(Some(&ts)));
-    }
-
-    #[test]
-    fn format_age_seconds() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
-        let ts = prost_types::Timestamp {
-            seconds: now.as_secs() as i64 - 30,
-            nanos: 0,
-        };
-        assert_eq!("30s ago", format_age(Some(&ts)));
-    }
-
-    #[test]
-    fn format_age_minutes() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
-        let ts = prost_types::Timestamp {
-            seconds: now.as_secs() as i64 - 64,
-            nanos: 0,
-        };
-        assert_eq!("1m4s ago", format_age(Some(&ts)));
-    }
-
-    #[test]
-    fn format_age_hours() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
-        let ts = prost_types::Timestamp {
-            seconds: now.as_secs() as i64 - (2 * 3600 + 3 * 60),
-            nanos: 0,
-        };
-        assert_eq!("2h3m ago", format_age(Some(&ts)));
-    }
 
     #[test]
     fn kind_display_all_variants() {
@@ -283,25 +208,21 @@ mod test {
 
     #[test]
     fn last_seen_cell_external_shows_age() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
+        let now = SystemTime::now().duration_since(std::time::UNIX_EPOCH).unwrap();
         let ts = prost_types::Timestamp {
             seconds: now.as_secs() as i64 - 5,
             nanos: 0,
         };
-        assert_eq!("5s ago", last_seen_cell(BackendKind::External, Some(&ts)));
+        assert_ne!("\u{2014}", last_seen_cell(BackendKind::External, Some(&ts)));
     }
 
     #[test]
     fn last_seen_cell_unspecified_shows_age() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
+        let now = SystemTime::now().duration_since(std::time::UNIX_EPOCH).unwrap();
         let ts = prost_types::Timestamp {
             seconds: now.as_secs() as i64 - 5,
             nanos: 0,
         };
-        assert_eq!("5s ago", last_seen_cell(BackendKind::Unspecified, Some(&ts)));
+        assert_ne!("\u{2014}", last_seen_cell(BackendKind::Unspecified, Some(&ts)));
     }
 }
