+++ operators/decap/cli/Cargo.toml
@@ -15,7 +15,6 @@ readinesspb = { path = "../../../common/rust/readinesspb", version = "0.1", pack
 clap = { version = "4.5", features = ["derive"] }
 clap_complete = { version = "4.5", features = ["unstable-dynamic"] }
 prost = "0.13"
-prost-types = "0.13"
 log = "0.4"
 colored = "3"
 tabled = { version = "0.18", features = ["ansi"] }
