+++ operators/decap/cli/src/main.rs
@@ -4,6 +4,7 @@
 //! and reports per-scope readiness state.
 
 use core::fmt::{self, Display, Formatter};
+use std::time::SystemTime;
 
 use clap::{ArgAction, CommandFactory, Parser};
 use clap_complete::CompleteEnv;
@@ -20,6 +21,7 @@ use tonic::codec::CompressionEncoding;
 use ync::{
     client::{ConnectionArgs, LayeredChannel, Service},
     errors::Error,
+    humanfmt,
     output::{self, CommonFormat},
 };
 
@@ -236,43 +238,12 @@ impl From<&readinesspb::pb::Scope> for ReadinessRow {
         Self {
             scope: scope.name.clone(),
             state: state_cell.to_string(),
-            last_transition: format_age(scope.last_transition_time.as_ref()),
-            observed: format_age(scope.observed_at.as_ref()),
+            last_transition: humanfmt::format_age(scope.last_transition_time.as_ref(), SystemTime::now())
+                .unwrap_or_else(|| "-".to_string()),
+            observed: humanfmt::format_age(scope.observed_at.as_ref(), SystemTime::now())
+                .unwrap_or_else(|| "-".to_string()),
             reasons,
         }
-    }
-}
-
-/// Formats a `prost_types::Timestamp` as a human-readable relative age.
-///
-/// Returns `"-"` when `ts` is `None` or the zero sentinel
-/// (`seconds == 0 && nanos == 0`). Otherwise formats as `Xs ago`,
-/// `XmYs ago`, or `XhYm ago` depending on magnitude.
-pub fn format_age(ts: Option<&prost_types::Timestamp>) -> String {
-    let ts = match ts {
-        Some(ts) if ts.seconds != 0 || ts.nanos != 0 => ts,
-        _ => return "-".to_string(),
-    };
-
-    let now = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default();
-
-    let ts_secs = ts.seconds.max(0) as u64;
-    let now_secs = now.as_secs();
-
-    let secs = now_secs.saturating_sub(ts_secs);
-
-    if secs < 60 {
-        format!("{secs}s ago")
-    } else if secs < 3600 {
-        let minutes = secs / 60;
-        let remainder = secs % 60;
-        format!("{minutes}m{remainder}s ago")
-    } else {
-        let hours = secs / 3600;
-        let minutes = (secs % 3600) / 60;
-        format!("{hours}h{minutes}m ago")
     }
 }
 
@@ -291,56 +262,4 @@ fn print_readiness_table(rows: Vec<ReadinessRow>) {
 
     ync::display::fit_terminal_width(&mut table);
     println!("{table}");
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn format_age_none_returns_dash() {
-        assert_eq!("-", format_age(None));
-    }
-
-    #[test]
-    fn format_age_zero_sentinel_returns_dash() {
-        let ts = prost_types::Timestamp { seconds: 0, nanos: 0 };
-        assert_eq!("-", format_age(Some(&ts)));
-    }
-
-    #[test]
-    fn format_age_seconds() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
-        let ts = prost_types::Timestamp {
-            seconds: now.as_secs() as i64 - 30,
-            nanos: 0,
-        };
-        assert_eq!("30s ago", format_age(Some(&ts)));
-    }
-
-    #[test]
-    fn format_age_minutes() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
-        let ts = prost_types::Timestamp {
-            seconds: now.as_secs() as i64 - 64,
-            nanos: 0,
-        };
-        assert_eq!("1m4s ago", format_age(Some(&ts)));
-    }
-
-    #[test]
-    fn format_age_hours() {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap();
-        let ts = prost_types::Timestamp {
-            seconds: now.as_secs() as i64 - (2 * 3600 + 3 * 60),
-            nanos: 0,
-        };
-        assert_eq!("2h3m ago", format_age(Some(&ts)));
-    }
 }
