+++ operators/pipeline/cli/Cargo.toml
@@ -13,7 +13,6 @@ ync = { path = "../../../cli/core", version = "0.1", package = "yanet-cli" }
 clap = { version = "4.5", features = ["derive"] }
 clap_complete = { version = "4.5", features = ["unstable-dynamic"] }
 prost = "0.13"
-prost-types = "0.13"
 log = "0.4"
 colored = "3"
 serde = { version = "1", features = ["derive"] }
