+++ operators/route/cli/route/Cargo.toml
@@ -17,7 +17,6 @@ readinesspb = { path = "../../../../common/rust/readinesspb", version = "0.1", p
 clap = { version = "4.5", features = ["derive"] }
 clap_complete = { version = "4.5", features = ["unstable-dynamic"] }
 prost = "0.13"
-prost-types = "0.13"
 log = "0.4"
 netip = "0.3"
 colored = "3"
